+++ crate/server/src/core/operations/mod.rs
@@ -76,5 +76,7 @@ pub(crate) use rng_seed::rng_seed;
 pub(crate) use set_attribute::set_attribute;
 pub(crate) use sign::sign;
 pub(crate) use signature_verify::signature_verify;
-pub(crate) use state_utils::get_effective_state;
+pub(crate) use state_utils::{
+    get_effective_state, is_user_authorized_for_operation, select_unique_key_for_operation,
+};
 pub(crate) use validate::validate_operation;